Added the last match attribute to all the brackets
Used the subsequent games lists to determine the last matches in
each of the brackets which will be useful for traversal from the
root of the tree. The function was beginning to get horribly long,
so i added step comments to figure out what is going on at each
for loop.

Added some classes that will be used for the rendering of the
brackets, namely the PlayerSlot class.