Implement initial drawing for bracket1
Fixed a canvas resize issue where it was not resizing to fit the
window properly. I implemented a few recursive functions within
the redraw function that draw most of the first tournement bracket.
The winners are currently drawn directly on the matches, so the next
step will be to set up a step by step functionality on the mouse
click that slowly reveals each winner one by one. Also the lines
connecting the matches still need to be drawn. Baby steps.