Add connecting lines between matches and slots
Implemented the drawing all the lines connecting matches and
player slots. The each match is responsible for drawing the
straight lines on either side of itself as well as the the
two lines branching of the end of that line connecting to its
children. The Player slot is only responsible for the one line
on it's left or right.