Add specific logic to handle loser brackets
The drawing on the loser brackets had a lot of dangling player slots that
would be drawn on top of because matches too precedence. There
now some logic to determine how to render a backet. In the case
of bracket1, the left and right branches are mostly equally balanced
in terms of depth from the root, so it makes sense to render half
the games on the left, and half on the right.

Since Bracket2 and bracket3 will have a depth of 1 on the left and
about 9 on the right, it makes sense to render the first match on
the left and proceed to the right. After that was implemented, I
fixed the spacing for matches and slots that overlapped each other
in bracket two.