Created a match page and matches can be revealed
I coded quite a bit before realizing i hadn't pushed in hours.
The matches now do not display their winner by default. I instead
created a "page" which is loaded upon clicking on one of the matches.
This page shows who is facing who, and displays the log location. At
that point you will find the gamelog and play it in the visualizer.
When the game is completed you click the reveal button which makes
the match winner visible in the bracket.

Next, I should make all the PlayerSlots generated from matches
that do not have any previous games (essentially all of the loser
slots in the other brackets) not visible by default as well. When
the result of a match is made visible, it should also set the
losing PlayerSlots in loser brackets to visible as well.