Fix a bug where revealing one match reveal others
The onclick event for the main canvas was storing all the events
handed to it, and when you called the on reveal function, it would
call all numerous copies of the matchPageClick handler where each
one accessed the outsides closures matched and revealed all of them
in turn. Modified such that a change of page requires the reassignment
of the click handler. Thank you jquery.