Fix bug where revealed match in 3rd bracket threw
It was just a type, It was looking for a match that it had just
determined was in bracket2 in bracket3. I didn't notice the throw
so it looked like the action of clicking the button was not doing
anything, and so i was checking that logic. Also changed a few
lines where it was storing the previous bracket function, because
quick switches was having the previous draw function change
before assigning to the variable, resulting in the back button
returning to another match page.